+++ AK/Format.cpp
@@ -26,6 +26,21 @@
 
 namespace AK {
 
+class FormatParser : public GenericLexer {
+public:
+    struct FormatSpecifier {
+        StringView flags;
+        size_t index;
+    };
+
+    explicit FormatParser(StringView input);
+
+    StringView consume_literal();
+    bool consume_number(size_t& value);
+    bool consume_specifier(FormatSpecifier& specifier);
+    bool consume_replacement_field(size_t& index);
+};
+
 namespace {
 
 static constexpr size_t use_next_index = NumericLimits<size_t>::max();
